Linux Support :D
Edited package.json so it is no longer forcing Electron to build a Windows Executable. Also changed controls so anything that requires pressing the alt key will now require pressing the Shift key. This is because Linux (maybe only Ubuntu and derivatives?) uses alt and clicking and dragging to move the window around. All control hints for actions using the Alt key have been updated to reflect this change.